Refactoring of crypto libraries
