+++ packages/skills/weather_station/skill.yaml
@@ -14,10 +14,7 @@ handlers:
       class_name: MyWeatherHandler
       args:
         foo: bar
-tasks:
-  - task:
-      class_name: MyWeatherTask
-      args:
-        foo: bar
+tasks: []
 shared_classes: []
 protocols: ['fipa']
+ledgers: []
